Refactor taskui and move buttons away
