+++ web/src/components/common/TaskUI.tsx
@@ -9,7 +9,8 @@ import {
   Tooltip,
   Text,
   Switch,
-  Spinner
+  Spinner,
+  Button
 } from '@chakra-ui/react'
 import React, { forwardRef, useCallback, useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
@@ -18,14 +19,12 @@ import AbortTaskButton from './AbortTaskButton'
 import { ChatSection } from './Chat'
 import { HiOutlineRefresh } from 'react-icons/hi'
 import { BiScreenshot, BiPaperclip } from 'react-icons/bi'
-import { HiMiniSparkles } from "react-icons/hi2";
 import chat from '../../chat/chat'
 import _ from 'lodash'
 import { abortPlan, startNewThread } from '../../state/chat/reducer'
 import { resetThumbnails, setInstructions as setTaskInstructions } from '../../state/thumbnails/reducer'
 import { setSuggestQueries } from '../../state/settings/reducer'
 import { RootState } from '../../state/store'
-import {  Button, Flex } from '@chakra-ui/react';
 import { getSuggestions } from '../../helpers/LLM/remote'
 import { Thumbnails } from './Thumbnails'
 import { UserConfirmation } from './UserConfirmation'
@@ -46,62 +45,9 @@ import { capture } from '../../helpers/screenCapture/extensionCapture'
 import { addThumbnail } from '../../state/thumbnails/reducer'
 import { Coordinates, startSelection } from '../../helpers/Selection'
 import { ImageContext } from '../../state/chat/types'
+import { QuickActionButton } from './QuickActionButton'
+import { ChatSuggestions } from './ChatSuggestions'
 
-
-interface ChatSuggestionsProps {
-  suggestQueries: boolean;
-  toggleSuggestions: (value: boolean) => void;
-  suggestions: string[];
-  onSuggestionClick: (suggestion: string) => void;
-}
-
-const ChatSuggestions: React.FC<ChatSuggestionsProps> = ({ suggestQueries, toggleSuggestions, suggestions, onSuggestionClick }) => {
-  return (
-    <Flex wrap="wrap" gap={2}>
-      <HStack justifyContent={"space-between"} width={"100%"}>
-        <HStack color="minusxGreen.500">
-          <HiMiniSparkles/>
-          <Text fontSize="sm" fontWeight={"bold"}>Suggestions</Text>
-        </HStack>
-        <HStack marginTop={0}>
-          <Switch color={"minusxBW.800"} colorScheme='minusxGreen' size={"sm"} isChecked={suggestQueries} onChange={(e) => toggleSuggestions(e.target.checked)} />
-        </HStack>
-      </HStack>
-
-      {suggestQueries && suggestions.length === 0 && (
-        <HStack justifyContent={"center"} width={"100%"}><Spinner size="xs" color="minusxGreen.500" /></HStack>
-      )}
-
-      {suggestQueries && suggestions.map((suggestion, index) => (
-        <Button
-          key={index}
-          onClick={() => onSuggestionClick(suggestion)}
-          size="sm"
-          colorScheme="blue"
-          variant="outline"
-          borderRadius="5px"
-          textAlign={"left"}
-          p={2}
-          style={{
-            whiteSpace: "normal",
-            wordWrap: "break-word",
-            height: "auto"
-          }}
-        >
-          {suggestion}
-        </Button>
-      ))}
-    </Flex>
-  );
-};
-
-const tooLongTooltip = (
-  <span>
-    Your message history is too long. 
-    <br/>
-     Try resetting.
-  </span>
-)
 
 const TaskUI = forwardRef<HTMLTextAreaElement>((_props, ref) => {
   const currentTool = useSelector((state: RootState) => state.settings.iframeInfo.tool)
@@ -118,6 +64,7 @@ const TaskUI = forwardRef<HTMLTextAreaElement>((_props, ref) => {
   const userConfirmation = activeThread.userConfirmation
   const dispatch = useDispatch()
   const taskInProgress = !(activeThread.status == 'FINISHED')
+  const isDevToolsOpen = useSelector((state: RootState) => state.settings.isDevToolsOpen)
 
   const debouncedSetInstruction = useCallback(
     _.debounce((instructions) => dispatch(setTaskInstructions(instructions)), 500),
@@ -221,16 +168,6 @@ const TaskUI = forwardRef<HTMLTextAreaElement>((_props, ref) => {
       runTask()
     }
   }
-  let resetMessageHistoryButton = (<IconButton
-    isRound={true}
-    onClick={clearMessages}
-    variant="ghost"
-    colorScheme="minusxGreen"
-    aria-label="Refresh"
-    size={'sm'}
-    icon={<Icon as={HiOutlineRefresh} boxSize={5} />}
-    disabled={messages.length === 0 || taskInProgress}
-    />)
 
   return (
     <VStack
@@ -358,29 +295,9 @@ const TaskUI = forwardRef<HTMLTextAreaElement>((_props, ref) => {
           <HStack position={"absolute"} bottom={0} width={"100%"} p={2}>
             <HStack justify={"space-between"}  width={"100%"}>
               <HStack gap={0}>
-                <Tooltip hasArrow label="Select & Ask" placement='right' borderRadius={5} openDelay={500}>
-                  <IconButton
-                    variant={'ghost'}
-                    colorScheme="minusxGreen"
-                    aria-label="Selection"
-                    size={'sm'}
-                    onClick={handleSnapClick}
-                    icon={<Icon as={BiPaperclip} boxSize={5} />}
-                  />
-                </Tooltip>
-                <Tooltip hasArrow label="Select & Ask" placement='right' borderRadius={5} openDelay={500}>
-                  <IconButton
-                    variant={'ghost'}
-                    colorScheme="minusxGreen"
-                    aria-label="Selection"
-                    size={'sm'}
-                    onClick={handleSnapClick}
-                    icon={<Icon as={BiScreenshot} boxSize={5} />}
-                  />
-                </Tooltip>
-                <Tooltip hasArrow label="Clear Chat" placement='right' borderRadius={5} openDelay={500}>
-                  {resetMessageHistoryButton}
-                </Tooltip>
+                <QuickActionButton tooltip="Add Context (Coming Soon!)" onclickFn={handleSnapClick} icon={BiPaperclip} isDisabled={true}/>
+                <QuickActionButton tooltip="Select & Ask" onclickFn={handleSnapClick} icon={BiScreenshot} isDisabled={false}/>
+                <QuickActionButton tooltip="Clear Chat" onclickFn={clearMessages} icon={HiOutlineRefresh} isDisabled={false}/>
               </HStack>
               <HStack>
                 {
